# EasyTrufi.Api/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Backend EasyTrufi API",
    "description": "Documentación de la API de EasyTrufi - .NET 8",
    "contact": {
      "name": "Equipo de Desarrollo UCB",
      "email": "carlos.camacho.h@ucb.edu.bo"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Driver": {
      "get": {
        "tags": [
          "Driver"
        ],
        "summary": "Obtiene una lista de conductores con soporte para paginación y filtros.",
        "parameters": [
          {
            "in": "query",
            "name": "Cedula",
            "type": "string"
          },
          {
            "in": "query",
            "name": "FullName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Active",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Driver"
        ],
        "summary": "Inserta un nuevo conductor en el sistema.",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "DTO con la información del conductor a insertar.",
            "schema": {
              "$ref": "#/definitions/DriverDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Driver/{id}": {
      "get": {
        "tags": [
          "Driver"
        ],
        "summary": "Obtiene un conductor específico por su identificador.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Identificador único del conductor.",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Driver"
        ],
        "summary": "Actualiza la información de un conductor existente.",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Identificador único del conductor a actualizar.",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "DTO con la nueva información del conductor.",
            "schema": {
              "$ref": "#/definitions/DriverDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Driver"
        ],
        "summary": "Elimina un conductor del sistema.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Identificador único del conductor a eliminar.",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "DriverDTO": {
      "type": "object",
      "properties": {
        "cedula": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "active": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
  }
}